road joining to get roads with resurfacing
The road selection by resurfacing is complete. This is somewhat hard coded, so a different borough will need some work. The roads are also not clipped, but just joined by name. This joining also accounts for duplicate roads. Will need to come up with a way to deal with duplicate roads by crossroad intersection.