a total mess, about to use geocode
The column sorting pivotting separating out, joining etc is in a good place, however the geocode is a mess. Finding street intersections is difficult in OSM and costs in Google. This attempts to bridge the gap using OSM to find the streets, then get the intersection of the two streets as a point (if not touching then snap), then route them